Simplified card naming
