+++ example/lib/screens/component/card/card_component_screen.dart
@@ -123,7 +123,7 @@ class _CardComponentScreenState extends State<CardComponentScreen> {
   ) {
     final paymentMethod = extractPaymentMethod(session.paymentMethodsJson);
 
-    return AdyenCardComponentWidget(
+    return AdyenCardComponent(
       configuration: cardComponentConfiguration,
       componentPaymentFlow: CardComponentSessionFlow(
         session: session,
