+++ example/lib/screens/component/card/card_component_scrollable_screen.dart
@@ -63,7 +63,7 @@ class CardComponentScrollableScreen extends StatelessWidget {
       cardConfiguration: const CardConfiguration(),
     );
 
-    return AdyenCardComponentWidget(
+    return AdyenCardComponent(
       configuration: cardComponentConfiguration,
       componentPaymentFlow: CardComponentAdvancedFlow(
         paymentMethod: paymentMethod,
